adding keyOptions
