+++ configdemo/src/App.jsx
@@ -27,9 +27,11 @@ function App() {
         <div className="field-modules-container">
           {isConfigActive && (
             <>
-              <FieldModules onDropdownChange={handleDropdownChange} />
+              {/* <FieldModules onDropdownChange={handleDropdownChange} />
               <FieldSelected selectedDropdowns={selectedDropdowns} 
-               setSelectedDropdowns={setSelectedDropdowns} />
+               setSelectedDropdowns={setSelectedDropdowns} /> */}
+               <AddConfig onDropdownChange={handleDropdownChange} />
+               <SelectConfig selectedDropdowns={selectedDropdowns} />
             </>
           )}
         </div>
