+++ configdemo/src/core/KeyOptions.js
@@ -1,0 +1,110 @@
+const hotelOptions = [
+  "code",
+  "name",
+  "hotelId",
+  "pms",
+  "fetchFromDb",
+  "PWAEnabled",
+  "saveToDb",
+  "modules",
+  "button_check_in",
+  "button_checkout",
+  "button_door_key",
+  "backgroundType",
+  "roomAutoAllocation",
+  "details",
+  "label",
+  "type",
+  "validations",
+  "isDisabled",
+  "isActive",
+  "title",
+  "documentScanType",
+  "faceMatchType",
+  "keyCardFaceMatchFlow",
+  "limitKeyCardLimit",
+  "addGuestMandatory",
+  "showDocumentExpiry",
+  "totalAddGuestCount",
+  "description",
+  "submodules",
+  "required",
+  "isAddCard",
+  "pricePerNight",
+  "subNote",
+  "updates",
+  "termsAndCondition",
+  "description1",
+  "description2",
+  "paymentType",
+  "keyVerificationFlow",
+  "sentInvoice",
+  "feedback",
+  "faceMatchFlow",
+  "feedBackURL1",
+  "feedBackURL2",
+  "feedBackURL1Label",
+  "feedBackURL2Label",
+  "paymentSuccessful",
+  "module",
+  "moduleTitle",
+];
+
+export { hotelOptions };
+
+const moduleOptions = [
+  "pullman-orchard-stage",
+  "Pullman Singapore Orchard Stage",
+  "989814e5-2fda-4ab8-b795-33f4d76f1866",
+  "OPERA",
+  "yes",
+  false,
+  "no",
+  "home",
+  "Begin your check-in",
+  "Checkout",
+  "Door Key",
+  "mp4",
+  "faceScan",
+  "bookingId",
+  true,
+  "confirmationId",
+  "Booking ID",
+  "Text",
+  "required",
+  "textField",
+  false,
+  true,
+  "lastname",
+  "Last Name",
+  "Room Number (In 4 Digits)",
+  "stayDetails",
+  "checkoutKey",
+  "Place Your Key",
+  "scanDoc",
+  "scan_api",
+  "incode",
+  "keyCard",
+  "cloud_api",
+  2,
+  "review",
+  "Identity Verification",
+  false,
+  4,
+  "accompanyGuest",
+  "Accompanying Guest",
+  "firstName",
+  "First Name",
+  "docNumber",
+  "Document Number",
+  "email",
+  "Card Confirmation",
+  "OperaTerminal",
+  "Incidental amount",
+  50,
+  "Card Authorization",
+  "Card Authorization Successful",
+  "Payment Successful",
+];
+
+export { moduleOptions };
